add context watcher guard as associated method
